requirements.txt
django-multiupload
django
huey
pyyaml
channels
channels_redis
